+++ src/main/scala/org/broadinstitute/hail/utils/package.scala
@@ -324,10 +324,10 @@ package object utils extends Logging
     }
   }
 
-  def getIteratorSize[T](iterator: Iterator[T]): Int = {
-    var count = 0
+  def getIteratorSize[T](iterator: Iterator[T]): Long = {
+    var count = 0L
     while (iterator.hasNext) {
-      count += 1
+      count += 1L
       iterator.next()
     }
     count
